update directory hierarchy
